# Reduce heap allocations in proc.ListPIDs()