A small improvement to proc.ListPIDs().

Benchmarks from my local dev system:

**Before**
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/ebpf-profiler/proc
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
BenchmarkListPIDs-20                5373            203274 ns/op           57888 B/op       1025 allocs/op
PASS
ok      go.opentelemetry.io/ebpf-profiler/proc  1.139s
```

**After**
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/ebpf-profiler/proc
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
BenchmarkListPIDs-20                7153            171458 ns/op           24855 B/op        614 allocs/op
PASS
ok      go.opentelemetry.io/ebpf-profiler/proc  1.642s
```