@fabled 2024-10-25
I am thinking if we could just completely drop `populatePIDs`? It is called only at startup, and the intention is to pre-populate all processes to ebpf. However, I think this is unnecessary, as the new-process-seen mechanism will report the ones that are actually busy. Just removing this would speed up startup and reduce memory usage everywhere as the sleeping processes are not processed.

@christos68k @florianl What do you think?