@rockdaboot 2024-10-25
> I think this is unnecessary, as the new-process-seen mechanism will report the ones that are actually busy. Just removing this would speed up startup and reduce memory usage everywhere as the sleeping processes are not processed.

Tbh, I wasn't aware of the lazy PID processing.
Removed the initial scanning for running processes.